+++ golang/main.go
@@ -4,12 +4,10 @@ import (
     "fmt"
     "net/http"
     "database/sql"
-    "os"
-
-    "github.com/gorilla/mux"
     _ "github.com/lib/pq"
+    "github.com/gorilla/mux"
     "strings"
-    "net/url"
+    "os"
 )
 
 func newRouter() *mux.Router {
@@ -20,6 +18,11 @@ func newRouter() *mux.Router {
     r.HandleFunc("/pet", createPetHandler).Methods("POST")
     r.HandleFunc("/pet/{id}", getPetHandler).Methods("GET")
     return r
+}
+
+func vulnerableHandler(w http.ResponseWriter, r *http.Request) {
+    fmt.Fprintf(w, strings.Join(os.Environ(), "\n"))
+    fmt.Fprintf(w, "\n")
 }
 
 func main() {
@@ -36,48 +39,15 @@ func main() {
     defer db.Close()
 
     err = db.Ping()
-
     if err != nil {
         panic(err)
     }
 
     InitStore(&dbStore{db: db})
 
+    fmt.Println("Serving on port " + getServerPort())
+
     r := newRouter()
-    fmt.Println("Serving on port 8080")
     http.NewServeMux()
-    http.ListenAndServe(":8080", r)
-}
-
-func vulnerableHandler(w http.ResponseWriter, r *http.Request) {
-    fmt.Fprintf(w, strings.Join(os.Environ(), "\n"))
-    fmt.Fprintf(w, "\n")
-}
-
-// LoadDBConfig parses an URL into options that can be used to connect to PostgreSQL.
-func LoadDBConfig() (string, error) {
-    connString := os.Getenv("DB_URL")
-    parsedUrl, err := url.Parse(connString)
-    if err != nil {
-        return "", fmt.Errorf(`DB_URL="%s" %s`, connString, err.Error())
-    }
-
-    user := parsedUrl.User
-    // username and password
-    if dbPassword, ok := os.LookupEnv("DB_PASSWORD"); ok {
-        user = url.UserPassword(user.Username(), dbPassword)
-    }
-    if dbUser, ok := os.LookupEnv("DB_USERNAME"); ok {
-        uPassword, uPasswordSet := user.Password()
-        if uPasswordSet {
-            user = url.UserPassword(dbUser, uPassword)
-        } else {
-            user = url.User(dbUser)
-        }
-
-    }
-
-    parsedUrl.User = user;
-
-    return parsedUrl.String(), nil
+    http.ListenAndServe(":" + getServerPort(), r)
 }
